+++ convocatorias/templates/contact_form.html
@@ -17,7 +17,7 @@
         <i class="fa-solid fa-book"></i>
         {{form.asunto}}
     </div>
-    <div class="contenedor-campo" id="comentarios">
+    <div class="contenedor-campo big-input" id="comentarios">
         <i class="fa-solid fa-pencil"></i>
         {{form.comentarios}}
     </div>
