+++ convocatorias/views.py
@@ -1,6 +1,11 @@
 from django.shortcuts import render, redirect
 from soporte.forms import SolicitudApoyoForm
+from .models import Modalidad
+
 # Create your views here.
+
+def obtener_becas(limite: int) -> list[Modalidad]:
+    return Modalidad.objects.all()[:limite]
 
 def index(request):
     if request.method == 'POST':
@@ -10,7 +15,7 @@ def index(request):
             return redirect('index')
     else:
         formulario_solicitud = SolicitudApoyoForm()
-    return render(request, 'index.html', {'form': formulario_solicitud})
+    return render(request, 'index.html', {'form': formulario_solicitud, "becas": obtener_becas(3)})
 
 def solicitud_de_apoyos(request):
     return render(request, 'solicitud_apoyo.html')
