+++ static/css/main.css
@@ -126,7 +126,7 @@ p {
 */
 
 .topbar-container {
-    background-image: url("../img/backgrounds/hero.jpg");
+    /* background-image: url("../img/backgrounds/hero.jpg"); */
     background-size: 100% 115%;
     background-repeat: no-repeat;
     background-position: center center;
@@ -251,24 +251,32 @@ p {
         gap: 20px;
         margin: 10px 0;
         padding: 20px;
-        align-items: center;
+        align-items: stretch;
+        height: 27rem;
     }
 }
 
 .becas .beca {
-    display: flex;
-    flex-direction: column;
+    display: grid;
+    grid-template-rows: 4rem 3rem 1.7rem 2.7rem 5.6rem 2.7rem;
+    gap: 0.8rem;
     align-items: center;
+    justify-items: center;
     border: solid 1px var(--beca-card-border-1);
     border-radius: 0.9rem;
-    flex-grow: 1;
     padding: 2rem 0.8rem;
     text-align: center;
     max-width: 13rem;
 }
 
-.becas .beca > *:not(:first-child):not(:last-child) {
-    margin: 0.8rem 0;
+@media (max-width: 768px) {
+    .becas > *:not(:first-child):not(:last-child) {
+        margin: 0.8rem 0;
+    }
+}
+
+.becas .beca h3 {
+    text-transform: uppercase;
 }
 
 @media (min-width: 768px) {
@@ -329,7 +337,7 @@ p {
     }
 }
 
-.support-sect {
+.form-style1 {
     border: solid 0.1rem var(--beca-card-border-1);
     border-radius: 0.2rem;
     margin-top: 1rem;
@@ -337,12 +345,12 @@ p {
     padding-top: 2rem;
 }
 
-.support-sect h2 {
+.form-style1 h2 {
     font-size: 2rem;
     color: var(--bg-regular-color);
 }
 
-.support-sect form {
+.form-style1 form {
     display: flex;
     flex-direction: column;
     width: 100%;
@@ -351,34 +359,34 @@ p {
 }
 
 @media (min-width: 768px) {
-    .support-sect form {
+    .form-style1 form {
         display: grid;
         grid-template-columns: repeat(2, 1fr);
         row-gap: 1.5rem;
     }
 
-    .support-sect form>.contenedor-campo[id="comentarios"] {
+    .form-style1 form .big-input {
         grid-column-end: span 2;
-    }
-
-    .support-sect form>.contenedor-campo[id="comentarios"] .regular-input {
-        min-height: 5rem;
     }
 }
 
-.support-sect form .contenedor-campo {
+.form-style1 form .big-input .regular-input {
+    min-height: 5rem;
+}
+
+.form-style1 form .contenedor-campo {
     display: flex;
     margin: 0.5rem 0.5rem;
 }
 
 @media (min-width: 768px) {
-    .support-sect form .contenedor-campo {
+    .form-style1 form .contenedor-campo {
         display: flex;
         margin: 0 0.5rem;
     }
 }
 
-.support-sect form i {
+.form-style1 form i {
     display: inline-block;
     color: white;
     background-color: var(--bg-regular-color);
@@ -388,18 +396,18 @@ p {
     border-radius: 0.3rem 0 0 0.3rem;
 }
 
-.support-sect form .contenedor-campo .regular-input {
+.form-style1 form .contenedor-campo .regular-input {
     flex-grow: 2;
     height: 2rem;
     border-radius: 0 0.3rem 0.3rem 0;
     padding-left: 0.5rem;
 }
 
-.support-sect form>.contenedor-campo:last-child .regular-input {
+.form-style1 form >.contenedor-campo:last-child .regular-input {
     height: 5rem;
 }
 
-.support-sect form button {
+.form-style1 form button {
     margin-left: 0.5rem;
 }
 
@@ -641,9 +649,6 @@ footer{
     
     section div h4{
         font-size: 20px;
-    }
-    section div {
-        flex-wrap: wrap;
     }
 }
 @media (min-width: 0) and (max-width: 1000px) {
